+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Display registration statistics in the BE in all cases (#2909)
 - Avoid accessing the deprecated `TYPO3_MODE` constant (#2906)
 - Upgrade the XLIFF files to XLIFF 1.2 (#2890)
 - Access superglobals directly (#2878)